pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "thunderousblushingprogrammer"
version = "0.1.0"
[tool.poetry.dependencies]
aiohttp = "^3.6"
python = "^3.7"
Pillow = "^7.1"
[tool.poetry.dev-dependencies]
